# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Smart Building API
  description: Sistema de controle inteligente de ar-condicionado
  version: 1.0.0
paths:
...
  /api/v1/auth/login:
    post:
      tags:
        - Auth
      summary: Autenticar e obter tokens JWT
      operationId: login_api_v1_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/refresh:
    post:
      tags:
        - Auth
      summary: Renovar access token
      operationId: refresh_api_v1_auth_refresh_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/logout:
    post:
      tags:
        - Auth
      summary: Revogar token (logout)
      operationId: logout_api_v1_auth_logout_post
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: Response Logout Api V1 Auth Logout Post
      security:
        - HTTPBearer: []
...
  /api/v1/alerts:
    get:
      tags:
        - Alerts
      summary: Listar alertas
      operationId: list_alerts_api_v1_alerts_get
      security:
        - HTTPBearer: []
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
            title: Page
        - name: size
          in: query
          required: false
          schema:
            type: integer
            maximum: 100
            minimum: 1
            default: 20
            title: Size
        - name: active_only
          in: query
          required: false
          schema:
            type: boolean
            description: Somente alertas não resolvidos
            default: false
            title: Active Only
          description: Somente alertas não resolvidos
        - name: device_id
          in: query
          required: false
          schema:
            anyOf:
              - type: string
                format: uuid
              - type: 'null'
            title: Device Id
        - name: severity
          in: query
          required: false
          schema:
            anyOf:
              - $ref: '#/components/schemas/AlertSeverity'
              - type: 'null'
            title: Severity
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: Response List Alerts Api V1 Alerts Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
        - Alerts
      summary: Criar alerta manualmente
      operationId: create_alert_api_v1_alerts_post
      security:
        - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlertCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/alerts/history:
    get:
      tags:
        - Alerts
      summary: Histórico de alertas
      operationId: get_alert_history_api_v1_alerts_history_get
      security:
        - HTTPBearer: []
      parameters:
        - name: days
          in: query
          required: false
          schema:
            type: integer
            maximum: 365
            minimum: 1
            default: 30
            title: Days
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
            title: Page
        - name: size
          in: query
          required: false
          schema:
            type: integer
            maximum: 200
            minimum: 1
            default: 50
            title: Size
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: Response Get Alert History Api V1 Alerts History Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  '/api/v1/alerts/{alert_id}':
    get:
      tags:
        - Alerts
      summary: Detalhes de um alerta
      operationId: get_alert_api_v1_alerts__alert_id__get
      security:
        - HTTPBearer: []
      parameters:
        - name: alert_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Alert Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  '/api/v1/alerts/{alert_id}/acknowledge':
    post:
      tags:
        - Alerts
      summary: Reconhecer alerta
      operationId: acknowledge_alert_api_v1_alerts__alert_id__acknowledge_post
      security:
        - HTTPBearer: []
      parameters:
        - name: alert_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Alert Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  '/api/v1/alerts/{alert_id}/resolve':
    post:
      tags:
        - Alerts
      summary: Resolver alerta
      operationId: resolve_alert_api_v1_alerts__alert_id__resolve_post
      security:
        - HTTPBearer: []
      parameters:
        - name: alert_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Alert Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/consumption:
    get:
      tags:
        - Consumption
      summary: Histórico de consumo energético
      operationId: get_consumption_api_v1_consumption_get
      security:
        - HTTPBearer: []
      parameters:
        - name: period
          in: query
          required: false
          schema:
            type: string
            pattern: ^(24h|7d|30d)$
            default: 24h
            title: Period
        - name: device_id
          in: query
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Device Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: Response Get Consumption Api V1 Consumption Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/consumption/summary:
    get:
      tags:
        - Consumption
      summary: Resumo de consumo por sala
      operationId: get_consumption_summary_api_v1_consumption_summary_get
      security:
        - HTTPBearer: []
      parameters:
        - name: period
          in: query
          required: false
          schema:
            type: string
            pattern: ^(24h|7d|30d)$
            default: 7d
            title: Period
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: >-
                  Response Get Consumption Summary Api V1 Consumption Summary
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/reports/consumption:
    get:
      tags:
        - Reports
      summary: Relatório de consumo
      operationId: get_consumption_report_api_v1_reports_consumption_get
      security:
        - HTTPBearer: []
      parameters:
        - name: period
          in: query
          required: false
          schema:
            type: string
            pattern: ^(24h|7d|30d)$
            default: 7d
            title: Period
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: Response Get Consumption Report Api V1 Reports Consumption Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/reports/alerts:
    get:
      tags:
        - Reports
      summary: Relatório de alertas
      operationId: get_alerts_report_api_v1_reports_alerts_get
      security:
        - HTTPBearer: []
      parameters:
        - name: period
          in: query
          required: false
          schema:
            type: string
            pattern: ^(24h|7d|30d)$
            default: 7d
            title: Period
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: Response Get Alerts Report Api V1 Reports Alerts Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
    AlertCreate:
      properties:
        device_id:
          type: string
          format: uuid
          title: Device Id
        alert_type:
          $ref: '#/components/schemas/AlertType'
        severity:
          $ref: '#/components/schemas/AlertSeverity'
        message:
          anyOf:
            - type: string
            - type: 'null'
          title: Message
      type: object
      required:
        - device_id
        - alert_type
        - severity
      title: AlertCreate
    AlertResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        device_id:
          type: string
          format: uuid
          title: Device Id
        alert_type:
          $ref: '#/components/schemas/AlertType'
        severity:
          $ref: '#/components/schemas/AlertSeverity'
        message:
          anyOf:
            - type: string
            - type: 'null'
          title: Message
        created_at:
          type: string
          format: date-time
          title: Created At
        acknowledged_at:
          anyOf:
            - type: string
              format: date-time
            - type: 'null'
          title: Acknowledged At
        resolved_at:
          anyOf:
            - type: string
              format: date-time
            - type: 'null'
          title: Resolved At
        is_active:
          type: boolean
          title: Is Active
          default: true
        links:
          items:
            $ref: '#/components/schemas/HateoasLink'
          type: array
          title: Links
      type: object
      required:
        - id
        - device_id
        - alert_type
        - severity
        - created_at
      title: AlertResponse
    AlertSeverity:
      type: string
      enum:
        - info
        - warning
        - critical
      title: AlertSeverity
    AlertType:
      type: string
      enum:
        - HIGH_TEMPERATURE
        - LOW_HUMIDITY
        - ANOMALY_DETECTED
        - DEVICE_OFFLINE
        - WINDOW_OPEN
        - CONSUMPTION_LIMIT
      title: AlertType
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    HateoasLink:
      properties:
        rel:
          type: string
          title: Rel
        href:
          type: string
          title: Href
        method:
          type: string
          title: Method
          default: GET
      type: object
      required:
        - rel
        - href
      title: HateoasLink
    LoginRequest:
      properties:
        email:
          type: string
          title: Email
        password:
          type: string
          title: Password
      type: object
      required:
        - email
        - password
      title: LoginRequest
...
    RefreshRequest:
      properties:
        refresh_token:
          type: string
          title: Refresh Token
      type: object
      required:
        - refresh_token
      title: RefreshRequest
...
    TokenResponse:
      properties:
        access_token:
          type: string
          title: Access Token
        refresh_token:
          type: string
          title: Refresh Token
        token_type:
          type: string
          title: Token Type
          default: bearer
        expires_in:
          type: integer
          title: Expires In
      type: object
      required:
        - access_token
        - refresh_token
        - expires_in
      title: TokenResponse
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
        - loc
        - msg
        - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
...